+++ packages/hyperliquid-composer/src/operations/spotMeta.ts
@@ -1,7 +1,7 @@
 import { Wallet } from 'ethers'
 
 import { HyperliquidClient } from '../signer'
-import { BaseInfoRequest, SpotMeta } from '../types'
+import { BaseInfoRequest, SpotMeta, SpotInfo } from '../types'
 
 export async function getSpotMeta(wallet: Wallet | null, isTestnet: boolean, logLevel: string, tokenIndex: string) {
     const action: BaseInfoRequest = {
@@ -15,5 +15,23 @@ export async function getSpotMeta(wallet: Wallet | null, isTestnet: boolean, log
     if (!token) {
         throw new Error(`Token ${tokenIndex} not found`)
     }
+    return token
+}
+
+export async function getHipTokenInfo(
+    wallet: Wallet | null,
+    isTestnet: boolean,
+    logLevel: string,
+    tokenAddress: string
+) {
+    const action: BaseInfoRequest = {
+        type: 'tokenDetails',
+        tokenId: tokenAddress,
+    }
+
+    const hyperliquidClient = new HyperliquidClient(isTestnet, logLevel)
+    const response = await hyperliquidClient.submitHyperliquidAction('/info', wallet, action)
+    const token = response as SpotInfo
+
     return token
 }