+++ examples/oft-hyperliquid/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @layerzerolabs/oft-hyperliquid-example
 
+## 0.0.4
+
+### Patch Changes
+
+- 1900486: docs typo cleanup
+
+## 0.0.3
+
+### Patch Changes
+
+- 690885c: get asset bridge address for any token and listing hip-1 token information - https://hyperliquid.gitbook.io/hyperliquid-docs/for-developers/api/info-endpoint/spot#retrieve-information-about-a-token
+
 ## 0.0.2
 
 ### Patch Changes