+++ packages/hyperliquid-composer/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @layerzerolabs/hyperliquid-composer
 
+## 0.0.8
+
+### Patch Changes
+
+- 1900486: docs typo cleanup
+
+## 0.0.7
+
+### Patch Changes
+
+- 690885c: get asset bridge address for any token and listing hip-1 token information - https://hyperliquid.gitbook.io/hyperliquid-docs/for-developers/api/info-endpoint/spot#retrieve-information-about-a-token
+
 ## 0.0.6
 
 ### Patch Changes